Adding a custom copy code component (#135)
* Adding a custom copy code component that allows us to intercept the highlighting

* Fixing bad css import

* Add 'cursor: pointer' to .copy-button in playground.css

Added cursor property to .copy-button CSS class in Playground component to enhance user experience. This will change the cursor icon to a pointer on hover over the 'copy' button, indicating that the button is clickable.

* Change CSS file import for code snippet component

The CSS import in the copy-code-snippet.tsx component has been updated. "highlight.css" is replaced with "prism-default.css". This improves syntax highlighting and the readability of the code. The name change better represents the file's functionality.