Refactor layout component and introduce SideNav component
Replaced the code in the layout component with a new SideNav component to improve code readability and maintainability. The SideNav component has its own state and effects for handling routing and active navigation items, which simplifies the layout component. Styles related to the side menu were refactored into a dedicated module for SideNav. This change allows for more granular control of navigation related logic