adding in the principles as static pages (#125)
* adding in the principles as static pages

* fixing the workflow ignore

* Fixed footer placement
using flex-box css rule to grow body content

* Refactor layout component and introduce SideNav component

Replaced the code in the layout component with a new SideNav component to improve code readability and maintainability. The SideNav component has its own state and effects for handling routing and active navigation items, which simplifies the layout component. Styles related to the side menu were refactored into a dedicated module for SideNav. This change allows for more granular control of navigation related logic

* Update website content, enhance usability

Updated content on the website to provide more thorough information and better user experience. The modifications include extending sections with more detailed descriptions and explanations, restructuring the content for better navigation, and replacing some terminology for improved readability.

* Adding a fix for the Table of Contents to make it styled properly and displayed correctly.

* Correct typo in 'Content design' section of the document.

A small grammatical error was present in the content design document. The sentence "Content design make things easier for users to understand." has been corrected to "Content design makes things easier for users to understand." This change ensures the readability and correctness of the document.

* Fix typo in principles

Co-authored-by: Greg Williams <51328760+UKHOGreg@users.noreply.github.com>

---------

Co-authored-by: lewis stevens <lewis.stevens1998@hotmail.com>
Co-authored-by: Greg Williams <51328760+UKHOGreg@users.noreply.github.com>